# Fix TurboRouter hash normalization for backward compatibility