## Issue Summary

Fixes a critical issue where TurboRouter queries registered with raw hashes (such as those from external databases like PrintOptim Backend) wouldn't match FraiseQL's normalized hash calculation, preventing turbo router activation.

**Impact**: Queries would always use `mode: "normal"` (~140ms) instead of `mode: "turbo"` (<20ms), significantly degrading performance for cached queries.

## Root Cause

The issue was a hash mismatch between:
- **External systems**: Calculate raw query hashes directly from query strings
- **FraiseQL**: Uses normalized hash calculation that collapses whitespace differently

Example from PrintOptim Backend:
- Raw hash: `859f5d3b94c4c1add28a74674c83d6b49cc4406c1292e21822d4ca3beb76d269`
- FraiseQL normalized hash: `dbd3faa3f5769a537420c5fe6215474e7c00bc9bc4f4ead7b5cf4fd9284cf217`

## Solution

### 1. Enhanced Hash Normalization (`hash_query()`)
- Improved regex-based whitespace normalization
- Better handling of GraphQL syntax characters
- Ensures queries like `query{user{id}}` and `query { user { id } }` hash identically

### 2. Backward Compatibility Methods
- **`hash_query_raw()`**: Computes hash directly from raw query string
- **`register_with_raw_hash()`**: Allows registration with pre-computed database hashes
- **Enhanced `get()`**: Fallback lookup tries normalized hash first, then raw hash

### 3. Integration Support
- New example demonstrates integration with database-stored queries
- Complete test suite validates the fix workflow
- All existing functionality preserved

## Changes Made

### Core Changes (`src/fraiseql/fastapi/turbo.py`)
```python
# Enhanced normalization with better regex patterns
def hash_query(self, query: str) -> str:
    # Step 1: Remove comments
    # Step 2: Normalize all whitespace to single spaces  
    # Step 3: Remove spaces around GraphQL syntax characters
    # Step 4: Add back minimal spaces for consistency

# New backward compatibility methods
def hash_query_raw(self, query: str) -> str:
    """Generate hash without normalization for backward compatibility."""
    
def register_with_raw_hash(self, turbo_query: TurboQuery, raw_hash: str) -> str:
    """Register query with specific pre-computed hash."""

def get(self, query: str) -> TurboQuery | None:
    """Try normalized hash first, then raw hash fallback."""
```

### Testing (`tests/test_turbo_router_hash_issue.py`)
- Reproduces the original PrintOptim Backend issue
- Validates hash normalization behavior across edge cases  
- Tests backward compatibility scenarios
- Verifies complete fix workflow

### Integration Example (`examples/printoptim_backend_integration.py`)
- Demonstrates loading database-registered queries
- Shows proper usage of `register_with_raw_hash()`
- Validates end-to-end functionality

## Validation

✅ **All existing tests pass** (2861 passed, 1 skipped)  
✅ **New comprehensive test suite** validates the fix  
✅ **Integration example** demonstrates real-world usage  
✅ **Mode selector** correctly chooses TURBO mode  
✅ **Backward compatibility** preserved  

## Performance Impact

**Before Fix:**
- Queries with raw hashes: `mode: "normal"`, ~140ms response time
- Hash mismatch prevented turbo router activation

**After Fix:**
- Same queries: `mode: "turbo"`, <20ms response time  
- Automatic fallback ensures compatibility with all registration methods

## Migration Guide

For systems with database-stored raw hashes:

```python
# Before (wouldn't work)
registry.register(turbo_query)  # Uses normalized hash

# After (works with database hashes)
registry.register_with_raw_hash(turbo_query, db_stored_raw_hash)
```

## Breaking Changes

None. All changes are purely additive and maintain full backward compatibility.

## Related Issues

This fix resolves integration issues reported by PrintOptim Backend and potentially other systems that:
- Store pre-computed query hashes in databases
- Calculate hashes from raw query strings
- Expect turbo router activation for registered queries

## Testing

\`\`\`bash
# Run the new test suite
pytest tests/test_turbo_router_hash_issue.py -v

# Run the integration example
python examples/printoptim_backend_integration.py

# Verify all existing tests still pass
pytest tests/integration/caching/ -v
\`\`\`